melonmust-backend/backend/requirements.txt
fastapi
uvicorn
aiosmtplib
python-dotenv
resend
python-multipart
sqlalchemy
pyscopg2-binary
cryptography

